# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: tinycdi Workspace Platform — Public API
  version: "0.1.0"
  summary: |
    Public HTTP API of the tinycdi workspace platform (portal origin).
  description: |
    This API is the only interface end users and the portal talk to. It is served
    on the **portal origin**; interactive desktop traffic runs on a separate
    **session origin** reached only through the launch flow described under
    `POST /v1/workspaces/{workspaceId}/connections`.

    ## Multi-tenancy and authorization

    - Authentication is an OIDC-backed session cookie. The verified principal
      (issuer + subject + tenant) is the sole source of ownership: requests never
      carry an owner field and the server never trusts client-supplied identity.
    - Every resource is scoped to the caller's tenant. Regular users see only
      resources they own; tenant administrators see the tenant scope.
      Cross-tenant access returns `404 NOT_FOUND` (the resource is not
      distinguishable from non-existent), never a leak of existence.
    - Authorization to act on a workspace (view, start, stop, delete, connect,
      attach, purge) requires ownership or tenant-admin role; otherwise
      `403 FORBIDDEN` / `404 NOT_FOUND`.
    - End users never write Kubernetes custom resources. The API translates
      accepted intents into `Workspace` CRs through its own service account via a
      transactional outbox (see ADR 0002); Kubernetes RBAC denies direct CR
...

    ## Idempotency

...

...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /v1/workspaces/{workspaceId}/connections:
    post:
      operationId: createConnection
      summary: Issue a launch ticket for an interactive session
      description: |
        Issues a short-lived (60 s), single-use opaque **launch ticket** and a
        `launchUrl` on the **session origin**. The browser opens the session by
        POSTing the ticket to `launchUrl` (form POST on the session host); the
        gateway redeems it atomically through the broker and sets a host-only
        `Secure`/`HttpOnly`/`SameSite` session cookie, then redirects to a clean
        URL. The ticket never travels in a query string and is never logged; the
        server stores only its hash.

        Requirements: the workspace must be in phase `Ready` with
        `desiredState: Running`, otherwise `409 INVALID_STATE`. At most one
        interactive connection is authorized per workspace: if a live lease
        exists and `takeover` is not set, the call fails with
        `409 CONNECTION_IN_USE`. With `takeover: true` the previous session is
        revoked and fenced **before** the new ticket becomes usable.
      tags: [workspaces]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionRequest"
      responses:
        "201":
          description: Launch ticket issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaunchTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tcdi_session
      description: |
        OIDC-backed session cookie issued by the portal login flow
        (`Secure`, `HttpOnly`, `SameSite=Strict` on the portal origin).
        Required on every endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token obtained from the session bootstrap payload and echoed on
        every mutating request. Combined with `SameSite=Strict` cookies this
        blocks cross-site request forgery; a missing or mismatched token
        returns `403 CSRF_FAILED`. Required on all non-GET operations.
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Server-generated workspace identifier.
      schema:
        type: string
        pattern: "^ws_[A-Za-z0-9]{8,64}$"
        example: ws_01J4Z8KQ2M9XNBV3T7YH0R6D5E
...
  responses:
    BadRequest:
      description: |
        The request is malformed or violates the contract — `INVALID_REQUEST`
        (bad JSON, unknown field, missing confirmation nonce, invalid cursor).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired or invalid session — `UNAUTHENTICATED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated but not allowed — `FORBIDDEN` for authorization failures,
        `CSRF_FAILED` for a missing/mismatched CSRF token on a mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        Resource does not exist or is not visible to the caller's tenant —
        `NOT_FOUND`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: |
        State or idempotency conflict — `IDEMPOTENCY_CONFLICT` (key reused with
        a different body), `INVALID_STATE` (operation not valid in the current
        phase/record state), `CONNECTION_IN_USE` (a live interactive lease
        exists and `takeover` was not set), or `QUOTA_EXHAUSTED` (tenant/user
        quota has no headroom).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Rate limit exceeded — `RATE_LIMITED`. Honor `Retry-After`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error — `INTERNAL`. Safe to retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: |
        Transient dependency failure — `UNAVAILABLE` (store/broker
        unreachable). Safe to retry with backoff.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error code. Values and their `retryable`
        semantics:

        | code | HTTP | retryable | meaning |
        |---|---|---|---|
        | `UNAUTHENTICATED` | 401 | false | no/expired session; re-login |
        | `CSRF_FAILED` | 403 | false | missing/mismatched CSRF token; refresh session bootstrap |
        | `FORBIDDEN` | 403 | false | principal lacks permission on this resource |
        | `NOT_FOUND` | 404 | false | resource absent or outside caller's tenant |
        | `INVALID_REQUEST` | 400 | false | malformed body/params; fix the request |
        | `INVALID_TEMPLATE` | 422 | false | templateRef unknown, unpublished or disallowed |
        | `INVALID_STATE` | 409 | true | phase/record state forbids the op now; retry once it settles to a compatible state |
        | `IDEMPOTENCY_CONFLICT` | 409 | false | Idempotency-Key reused with a different body; generate a new key |
        | `QUOTA_EXHAUSTED` | 409 | false | tenant/user quota has no headroom; free resources or raise quota |
        | `CONNECTION_IN_USE` | 409 | false | a live interactive lease exists; pass `takeover: true` to replace it |
        | `RATE_LIMITED` | 429 | true | transient throttle; honor `Retry-After` |
        | `UNAVAILABLE` | 503 | true | transient dependency failure (store/broker); retry with backoff |
        | `INTERNAL` | 500 | true | unexpected server failure; safe to retry |

        Clients must treat unknown codes as `INTERNAL` (retryable: true).
      enum:
        - UNAUTHENTICATED
        - CSRF_FAILED
        - FORBIDDEN
        - NOT_FOUND
        - INVALID_REQUEST
        - INVALID_TEMPLATE
        - INVALID_STATE
        - IDEMPOTENCY_CONFLICT
        - QUOTA_EXHAUSTED
        - CONNECTION_IN_USE
        - RATE_LIMITED
        - UNAVAILABLE
        - INTERNAL
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, requestId]
      description: Uniform error body returned for every 4xx/5xx response.
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable detail; not part of the contract.
          example: "workspace ws_01J4Z8KQ2M9XNBV3T7YH0R6D5E is not in a startable phase"
        retryable:
          type: boolean
          description: |
            Whether retrying the same request may succeed without user action
            (see the `ErrorCode` table for per-code semantics).
        requestId:
          type: string
          description: Server-generated request ID for support/log correlation.
          example: req_01J4ZA1X0YJ8H6K0Q3MW1N5B2P
...
    CreateConnectionRequest:
      type: object
      additionalProperties: false
      properties:
        takeover:
          type: boolean
          default: false
          description: |
            When true and a live interactive lease exists, it is revoked and
            fenced before the new ticket becomes usable. When false a live
            lease yields `409 CONNECTION_IN_USE`.
    LaunchTicket:
      type: object
      additionalProperties: false
      required: [workspaceId, ticket, launchUrl, expiresAt]
      description: |
        One-time launch material for an interactive session. `ticket` is opaque,
        lives 60 s, is consumed on first redemption attempt (even if expired),
        and is stored server-side only as a hash. It is delivered in this
        response body only — the browser POSTs it to `launchUrl` on the session
        origin; it must never appear in a URL query string.
      properties:
        workspaceId:
          type: string
          pattern: "^ws_[A-Za-z0-9]{8,64}$"
        ticket:
          type: string
          minLength: 32
          description: Opaque single-use ticket (base64url, ~256 bits of entropy).
          example: "tkt_e3B0cmFuZG9tLWJ5dGVzLW9ubHk_0123abcd"
        launchUrl:
          type: string
          format: uri
          description: |
            Absolute URL on the **session origin** (a different registrable
            domain than the portal). The browser POSTs the ticket here; the
            gateway redeems it and redirects to a clean URL after setting the
            host-only session cookie.
          example: "https://session.example.invalid/v1/launch"
        expiresAt:
          type: string
          format: date-time
          description: Ticket expiry (issued + 60 s).
...